remove the useDefaultFormatters option and move the default dumpers to the default of the dumpers list
